Completed a passable SolarSystemNameGenerator
Added a couple of extra options to the constellations and added a list of other prefixes (gathered from various surveys). Tidied up the next() function and refactored the name generation into separate function and added generateNames(int amount) and clear() which returns a given amount of star system names and clears the current set of names respectively.